Beginning of radial symmetry
What: This commit has 2 functions in the file. One that finds key points in a photo and uses SIFT to do so and the other uses HoughCircles to find round objects in a photo. By using SIFT and HoughCircles, I am hoping to get them to find the radial symmetry in a photo.
Why: I am adding this in since I haven't been able to meet lately and want to show my work.
Testing: I found a couple photos of coins that I originally used to test it. I was able to get it to locate the edges of most coins but am having a little issues with dimes since they are small. I also used car photos to see if it could pick up the wheels as such in our paper. It did find the wheel but is drawing a huge circle over it instead of an oval (the car is usually facing diagonally). It is finding key points ok but not getting the circles well enough. 
Additional: I need to know how to find directions using the key points so I can get a better handle on it. Direction will also help with symmetry balance as well